requirements.txt
torch>=2.1.2

horde_sdk~=0.9.3
horde_safety~=0.2.3
hordelib~=2.8.1
horde_model_reference~=0.6.3

python-dotenv
ruamel.yaml
semver

python-Levenshtein

pydantic~=2.7.0
typing_extensions
requests
StrEnum
loguru

babel

PyYAML